refactor: isolate internal runtime and generation boundaries
