+++ internal/ui/ui_quote.go
@@ -41,13 +41,13 @@ func (pself *UiQuote) OnStart() {
 }
 
 func (pself *UiQuote) SetText(pos mathf.Vec2, size mathf.Vec2, msg, description string) {
-	mgr.UiMgr.SetScale(pself.GetId(), engine.UniformVec2(engine.WindowScale()))
+	engine.Managers().UiMgr.SetScale(pself.GetId(), engine.UniformVec2(engine.WindowScale()))
 	pos = engine.BridgeWorldToView(pos)
 	targetPos := pos.Sub(mathf.NewVec2(size.X, -size.Y))
-	mgr.UiMgr.SetGlobalPosition(pself.container.GetId(), ViewToUI(targetPos))
-	mgr.UiMgr.SetSize(pself.container.GetId(), size.Mulf(2))
-	mgr.UiMgr.SetText(pself.labelMsg.GetId(), msg)
-	mgr.UiMgr.SetText(pself.labelDes.GetId(), description)
+	engine.Managers().UiMgr.SetGlobalPosition(pself.container.GetId(), ViewToUI(targetPos))
+	engine.Managers().UiMgr.SetSize(pself.container.GetId(), size.Mulf(2))
+	engine.Managers().UiMgr.SetText(pself.labelMsg.GetId(), msg)
+	engine.Managers().UiMgr.SetText(pself.labelDes.GetId(), description)
 }
 
 func NewUiQuote() *UiQuote {
